Styled page with bootstrap
Styled page with bootstrap to make the data of each car more readable,
called through the IDs of the colour table to print the colour otherwise
we would only have the colour id. Also removed 2013 from the footer as
the current year is already pulled through; this did print 2013 twice
but has now been corrected.